Disable URL sessions in Jetty by default.
Having URL sessions on can result in leakage of a user's JSESSIONID,
which can then be used in an attack on the session; it can also result
in sending others a link that inadvertently provides access to the
original session. This is rarely needed and can be re-enabled in such
cases.